GNSS SDR now uses the unified Gnss_Synchro class to exchange  signal synchronization data between Acquisition, Tracking and Telemetry decoder blocks.
git-svn-id: https://svn.code.sf.net/p/gnss-sdr/code/trunk@147 64b25241-fba3-4117-9849-534c7e92360d

+++ src/algorithms/telemetry_decoder/gnuradio_blocks/gps_l1_ca_telemetry_decoder_cc.h
@@ -100,7 +100,7 @@ private:
   bool d_flag_frame_sync;
 
   // symbols
-  int d_symbol_accumulator;
+  double d_symbol_accumulator;
   short int d_symbol_accumulator_counter;
 
   //bits and frame
